@@ -12,27 +12,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.3.6",
-    "@angular/cdk": "^20.2.9",
-    "@angular/common": "20.0.6",
-    "@angular/compiler": "20.0.6",
-    "@angular/core": "20.0.6",
-    "@angular/forms": "20.0.6",
-    "@angular/material": "^20.2.9",
-    "@angular/platform-browser": "20.0.6",
-    "@angular/platform-browser-dynamic": "20.0.6",
-    "@angular/router": "20.0.6",
+    "@angular-architects/ngrx-toolkit": "^20.4.2",
+    "@angular/animations": "20.3.9",
+    "@angular/cdk": "20.2.11",
+    "@angular/common": "20.3.9",
+    "@angular/compiler": "20.3.9",
+    "@angular/core": "20.3.9",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "20.3.9",
+    "@angular/material": "20.2.11",
+    "@angular/platform-browser": "20.3.9",
+    "@angular/platform-browser-dynamic": "20.3.9",
+    "@angular/router": "20.3.9",
     "@fortawesome/angular-fontawesome": "^3.0.0",
     "@fortawesome/fontawesome-common-types": "^7.1.0",
     "@fortawesome/fontawesome-svg-core": "^7.0.1",
     "@fortawesome/free-brands-svg-icons": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^7.1.0",
-    "@fortawesome/pro-duotone-svg-icons": "^7.1.0",
     "@fortawesome/pro-light-svg-icons": "^7.1.0",
     "@fortawesome/pro-regular-svg-icons": "^7.1.0",
     "@fortawesome/pro-solid-svg-icons": "^7.1.0",
     "@fortawesome/pro-thin-svg-icons": "^7.1.0",
     "@fortawesome/sharp-solid-svg-icons": "^7.1.0",
+    "@ngrx/signals": "^20.1.0",
     "firebase": "^11.0.2",
     "firebase-tools": "^13.27.0",
     "punycode": "^2.3.1",
@@ -40,29 +42,30 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "20.0.5",
-    "@angular-devkit/core": "20.0.5",
-    "@angular-devkit/schematics": "20.0.5",
+    "@angular-devkit/build-angular": "20.3.8",
+    "@angular-devkit/core": "20.3.8",
+    "@angular-devkit/schematics": "20.3.8",
     "@angular/cli": "~20.0.0",
-    "@angular/compiler-cli": "20.0.6",
-    "@angular/language-service": "20.0.6",
+    "@angular/compiler-cli": "20.3.9",
+    "@angular/language-service": "20.3.9",
     "@eslint/js": "^9.8.0",
-    "@nx/angular": "21.2.1",
-    "@nx/cypress": "21.2.1",
-    "@nx/eslint": "21.2.1",
-    "@nx/eslint-plugin": "21.2.1",
-    "@nx/jest": "21.2.1",
-    "@nx/js": "21.2.1",
-    "@nx/web": "21.2.1",
-    "@nx/workspace": "21.2.1",
-    "@schematics/angular": "20.0.5",
+    "@ngrx/store": "^20.1.0",
+    "@nx/angular": "22.0.2",
+    "@nx/cypress": "22.0.2",
+    "@nx/eslint": "22.0.2",
+    "@nx/eslint-plugin": "22.0.2",
+    "@nx/jest": "22.0.2",
+    "@nx/js": "22.0.2",
+    "@nx/web": "22.0.2",
+    "@nx/workspace": "22.0.2",
+    "@schematics/angular": "20.3.8",
     "@swc-node/register": "~1.9.1",
     "@swc/core": "~1.5.7",
     "@swc/helpers": "~0.5.11",
     "@types/jest": "^29.5.12",
     "@types/node": "18.16.9",
     "@typescript-eslint/utils": "^8.29.0",
-    "angular-eslint": "20.0.0",
+    "angular-eslint": "20.3.0",
     "cypress": "^14.2.1",
     "eslint": "^9.8.0",
     "eslint-config-prettier": "^10.0.0",
@@ -70,12 +73,13 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-preset-angular": "14.6.0",
-    "nx": "21.2.1",
+    "node-fetch": "^3.3.2",
+    "nx": "22.0.2",
     "prettier": "^2.6.2",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "29.4.5",
     "ts-node": "10.9.1",
     "tslib": "^2.3.0",
-    "typescript": "5.8.3",
+    "typescript": "5.9.3",
     "typescript-eslint": "^8.19.0"
   },
   "packageManager": "yarn@4.9.1"
