@@ -2,12 +2,12 @@ import { Route } from '@angular/router';
 
 export const appRoutes: Route[] = [
   {
-    path: 'login',
-    loadChildren: () =>
-      import('@adrian-badilla/auth').then((m) => m.adrianBadillaUiAuthRoutes),
-  },
-  {
     path: '',
     loadChildren: () => import('@adrian-badilla/ui/landing-page/routes'),
   },
+  {
+    path: '',
+    loadChildren: () => import('@adrian-badilla/auth/routes'),
+  },
+  { path: '**', redirectTo: '' },
 ];
