@@ -1,5 +1,4 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { HeroComponent } from '../components/hero/hero.component';
 import { FooterComponent } from '../components/footer/footer.component';
 import { CardsComponent } from '../components/cards/cards.component';
@@ -9,17 +8,15 @@ import { ReviewsComponent } from '../components/reviews/reviews.component';
 import { AnnouncementComponent } from '../components/announcement/announcement.component';
 
 @Component({
-  standalone: true,
   selector: 'adrian-badilla-landing-page',
   imports: [
-    CommonModule,
     HeroComponent,
     FooterComponent,
     CardsComponent,
     GalleryComponent,
     TrainerInfoComponent,
     ReviewsComponent,
-    AnnouncementComponent
+    AnnouncementComponent,
   ],
   templateUrl: './adrian-badilla-ui-landing-page.component.html',
   styleUrl: './adrian-badilla-ui-landing-page.component.scss',
