@@ -5,10 +5,11 @@
 import { PrimaryAnimatedButtonComponent } from '../components/primary-animated-button/primary-animated-button.component';
 import { SecondaryAnimatedButtonComponent } from '../components/secondary-animated-button/secondary-animated-button.component';
 import { StatusMessageComponent } from '../components/status-message/status-message.component';
+import { StrengthMeterComponent } from '../components/strength-meter/strength-meter.component';
 
 export const COMPONENTS = [
   StatusMessageComponent,
-  // StrengthMeterComponent,
+  StrengthMeterComponent,
   PrimaryAnimatedButtonComponent,
   SecondaryAnimatedButtonComponent,
   // TertiaryAnimatedButtonComponent,
@@ -18,7 +19,7 @@ export const COMPONENTS = [
 ] as const;
 
 export * from '../components/status-message/status-message.component';
-// export * from '../components/strength-meter/strength-meter.component';
+export * from '../components/strength-meter/strength-meter.component';
 export * from '../components/primary-animated-button/primary-animated-button.component';
 // export * from '../components/tertiary-animated-button/tertiary-animated-button.component';
 export * from '../components/secondary-animated-button/secondary-animated-button.component';
